requirements.txt
fastapi
uvicorn
sqlalchemy
spacy
scikit-learn
transformers
pytesseract
pdfplumber
python-docx
pillow
pyMupdf
python-multipart
python-magic
torch
psycopg2-binary
boto3
python-dotenv